# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ecommerce REST API built using Go Lang, PSQL, REST API following Clean Architecture. Hosted with Ngnix, AWS EC2 and RDS",
        "title": "Ecommerce REST API",
        "contact": {
            "name": "Nadeem Fahad",
            "url": "https://github.com/Nadeem1815",
            "email": "nadeemf408@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/user/addcoupon/{couponid}": {
            "post": {
                "description": "User Can Add coupon To Cart",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "User Can Add coupon To Cart",
                "operationId": "add-coupon-to-cart",
                "parameters": [
                    {
                        "type": "string",
                        "description": "couponid",
                        "name": "couponid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    }
                }
            }
        },
...
        "/user/cart/{product_id}": {
            "post": {
                "description": "User Can add product To Cart",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "User Add Product To Cart By ProductId",
                "operationId": "user-add-to-cart-",
                "parameters": [
                    {
                        "type": "string",
                        "description": "product_item_id",
                        "name": "product_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    }
                }
            }
        },
        "/user/carts": {
            "get": {
                "description": "User Can View Cart and Total Amount",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "User Can View Cart and Total Amount",
                "operationId": "View-cart",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    }
                }
            }
        },
...
        "/user/remove/{product_id}": {
            "delete": {
                "description": "User Can Remove product To Cart",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "User Remove Product To Cart By ProductId",
                "operationId": "user-remove-to-cart-",
                "parameters": [
                    {
                        "type": "string",
                        "description": "product_item_id",
                        "name": "product_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_nadeem1815_premium-watch_pkg_utils_response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        }
    }
}